redundant

+++ src/main/java/at/hannibal2/skyhanni/features/inventory/wardrobe/WardrobeOverlay.kt
@@ -107,8 +107,8 @@ class WardrobeOverlay {
                     playerX,
                     playerY,
                     scale,
-                    eyesX.toFloat(),
-                    eyesY.toFloat(),
+                    eyesX,
+                    eyesY,
                     fakePlayer
                 )
